# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DocTrackr API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checklist:
    get:
      operationId: listChecklistItems
      tags: [checklist]
      summary: List all checklist items (compliance view)
      parameters:
        - name: personType
          in: query
          schema:
            type: string
            enum: [student, employee, all]
        - name: status
          in: query
          schema:
            type: string
            enum: [missing, valid, expiring_soon, due_today, overdue, all]
        - name: documentTypeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
        - name: peopleStatus
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
      responses:
        "200":
          description: Checklist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistItem"
  /checklist/{id}:
    get:
      operationId: getChecklistItem
      tags: [checklist]
      summary: Get checklist item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checklist item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistItem"
    patch:
      operationId: updateChecklistItem
      tags: [checklist]
      summary: Update checklist item (expiry date, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistItem"
  /checklist/bulk-update:
    patch:
      operationId: bulkUpdateChecklistItems
      tags: [checklist]
      summary: Bulk update multiple checklist items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkChecklistUpdate"
      responses:
        "200":
          description: Updated items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistItem"
  /checklist/{id}/send-reminder:
    post:
      operationId: sendManualReminder
      tags: [reminders]
      summary: Send manual reminder for a checklist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # Email Logs
...
components:
  schemas:
...
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    ChecklistItem:
      type: object
      properties:
        id:
          type: integer
        personType:
          type: string
          enum: [student, employee]
        personId:
          type: integer
        personName:
          type: string
        personEmail:
          type: ["string", "null"]
        personStatus:
          type: string
          enum: [active, inactive]
        documentTypeId:
          type: integer
        documentTypeName:
          type: string
        documentTypeCategory:
          type: string
        templateFormUrl:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [missing, valid, expiring_soon, due_today, overdue]
        notes:
          type: ["string", "null"]
        lastReminderSentAt:
          type: ["string", "null"]
        nextReminderDueAt:
          type: ["string", "null"]
        daysUntilExpiry:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - personType
        - personId
        - personName
        - personStatus
        - documentTypeId
        - documentTypeName
        - documentTypeCategory
        - status
        - createdAt
        - updatedAt
    ChecklistItemUpdate:
      type: object
      properties:
        expiryDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    BulkChecklistUpdate:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              expiryDate:
                type: ["string", "null"]
              notes:
                type: ["string", "null"]
            required:
              - id
      required:
        - items
...