# lib/api-zod/src/generated/types/checklistItemUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DocTrackr API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChecklistItemUpdate {
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  notes?: string | null;
}